requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.12.1
asyncpg==0.29.0

# Configuration Management
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# Payment Integrate
paystackapi==2.1.3

# Caching & Session Management
redis==5.0.1
hiredis==2.2.3

# HTTP Requests
httpx==0.25.1
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
uuid7==0.1.0

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
